+++ FluentStorage.GCP/Storage/GoogleCloudStore.cs
@@ -657,7 +657,7 @@ namespace FluentStorage.GCP.Storage {
 		/// <summary>
 		/// Fastest implementation of getting an object checksum using GCP metadata when available.
 		/// </summary>
-		public override async Task<StorageObjectHash> GetObjectChecksum(string fullPath, StorageHash hash = StorageHash.CRC32, CancellationToken cancellationToken = default) {
+		public override async Task<StorageObjectHash> GetObjectChecksum(string fullPath, StorageHash hash = StorageHash.MD5, CancellationToken cancellationToken = default) {
 
 			if (fullPath == null)
 				throw new ArgumentNullException(nameof(fullPath));