+++ FluentStorage/Enums/StorageExistsMode.cs
@@ -1,25 +1,26 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace FluentStorage.Enums {
+﻿namespace FluentStorage.Enums {
 	/// <summary>
 	/// Determines what to do when the destination file/object already exists.
 	/// </summary>
 	public enum StorageExistsMode {
 		/// <summary>
-		/// Throw an exception if the destination already exists.
+		/// Throw an exception if the object already exists.
 		/// </summary>
 		Throw,
 
 		/// <summary>
-		/// Skip the transfer if the destination already exists.
+		/// Skip the transfer if the object already exists.
 		/// </summary>
 		Skip,
 
 		/// <summary>
-		/// Replace the destination if it already exists.
+		/// Replace the object if it already exists.
 		/// </summary>
 		Overwrite,
+
+		/// <summary>
+		/// Replace the object if the object's length or checksum is different.
+		/// </summary>
+		OverwriteIfChanged,
 	}
 }