+++ FluentStorage/Enums/StorageUrlProtocol.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace FluentStorage.Enums {
+﻿namespace FluentStorage.Enums {
 	public enum StorageUrlProtocol {
 		Http = 1,
 		HttpAndHttps = 2,