+++ RELEASES.md
@@ -2,6 +2,19 @@
 
 **Please read the [Migration Guide](https://github.com/robinrodricks/FluentStorage/wiki/Migration-Guide) to help you migrate from older versions to FluentStorage 8!**
  
+#### FluentStorage 8.0.14
+ - **FluentStorage**
+   - New: `GetObjectChecksum` API to compute various types of hashes for an object across all providers
+   - New: `StorageObjectHash` object to hold a hash value and easily compare it against any local file or stream
+   - New: `StorageExistsMode.OverwriteIfChanged` mode to only overwrite files if they have changed (length/checksum checks)
+   - New: `UploadDirectory` supports the `OverwriteIfChanged` and only uploads files if the length/checksum mismatches
+   - New: `DownloadDirectory` supports the `OverwriteIfChanged` and only downloads files if the length/checksum mismatches
+   - Removed: `GetObjectMD5` is removed in favour of `GetObjectChecksum`
+ - **FluentStorage.GCP**
+   - New: Implement fast `GetObjectChecksum` by using `GetObjectAsync` API to get MD5 object hash
+ - **FluentStorage.FTP**
+   - New: Implement fast `GetObjectChecksum` by using native FTP commands
+ 
 #### FluentStorage 8.0.13
  - **FluentStorage**
    - New: Optimize `StoragePath.Combine` when combining only 2 path segments (most common use case)